Update vagrant provisioning script
Make some changes to the vagrant provisioning script to make it easier
to install gems with bundler.

The reason for upgrading for trusty is that the bundler package is a bit
broken on precise and gives permissions errors as it's an outdated
version. The trusty version of bundler is more up-to-date and works
nicely out of the box without any hoop jumping.

- Upgrade to a trusty64 base box
- Install the bundler package

+++ config/provision.sh
@@ -19,12 +19,14 @@ sudo DEBIAN_FRONTEND=noninteractive apt-get install -y \
   nodejs build-essential \
   mongodb-org \
   openjdk-6-jre elasticsearch \
-  git imagemagick graphicsmagick postfix redis-server
-
-bundle install --path vendor/bundle --binstubs vendor/bundle-bin
+  git imagemagick graphicsmagick postfix redis-server bundler
 
 grep -qG 'cd /vagrant' "$HOME/.bashrc" || echo "cd /vagrant" >> "$HOME/.bashrc"
 cd /vagrant
 
+bundle install
 make
-cp config/development.js-example config/development.js
+
+if [[ ! -f config/development.js ]]; then
+  cp config/development.js-example config/development.js
+fi
